Global:  Big update to support 64-bit Windows with MSYS2/MINGW64. The long type is 32-bits here, 64-bits everywhere else, so is replaced by appropriate types from stdint.h, usually uintptr_t which always has the same size as a pointer.

+++ wrspice/include/circuit.h
@@ -58,6 +58,7 @@ Authors: 1985 Thomas L. Quarles
 #define THREAD_SAFE_EVAL
 
 #include <math.h>
+#include <stdint.h>
 #include "ifdata.h"
 #ifdef WITH_THREADS
 #include <pthread.h>
@@ -161,7 +162,7 @@ struct sGENmodel;
 struct sGENinstance
 {
     // Hash table hooks.
-    unsigned long tab_key()             { return ((unsigned long)GENname); }
+    uintptr_t tab_key()                 { return ((uintptr_t)GENname); }
     sGENinstance *tab_next()            { return (GENnextTabInstance); }
     void set_tab_next(sGENinstance *n)  { GENnextTabInstance = n; }
